Added a screenshot with successful compilation.
OR chip was successfully implemented and tested build using 2 NAND and 1 NOT gate. Input a is passed for both inputs of NAND gate which results in NAND a and input b is passed through NOT resulting in negation of b - NOT b. Outputs of the above two chips NAND a & NOT b is passed through another NAND gate which gives the output that is same as OR gate. This was implemented using HDL and the outputs were tested successfully.